updateAssignedUsersForTemplate
